requirements.txt
prometheus-client>=0.20.0
paramiko>=3.4.0
PyYAML>=6.0.1
